Move BoxedSetter into its own module
